fix:메트릭 수정

+++ src/main/java/org/example/finalbe/domains/monitoring/dto/RackStatisticsDto.java
@@ -121,8 +121,8 @@ public class RackStatisticsDto {
         private Double maxUsage;
         private List<TopEquipment> topEquipments;
         private Integer equipmentCount;
-        private Long totalCapacityTB;
-        private Long usedCapacityTB;
+        private Double totalCapacityTB;
+        private Double usedCapacityTB;
     }
 
     @Getter
